requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.5
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
boto3==1.28.41
botocore==1.31.41
braceexpand==0.1.7
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.7.22
cffi==1.15.0
cfgv==3.4.0
charset-normalizer==3.2.0
chex==0.1.85
click==8.1.7
clip==1.0
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.1.4
confection==0.1.4
contextlib2==21.6.0
contexttimer==0.3.3
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.8
datasets==2.17.0
debugpy==1.6.7
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
distrax==0.1.2
distro==1.9.0
dm-tree==0.1.8
docker-pycreds==0.4.0
einops==0.7.0
emoji==2.8.0
entrypoints==0.4
etils==1.5.2
exceptiongroup==1.1.3
executing==1.2.0
fairscale==0.4.4
fastapi==0.109.2
fastjsonschema==2.18.1
ffmpy==0.3.2
filelock==3.12.2
flax==0.8.1
fonttools==4.42.1
frozenlist==1.4.1
fsspec==2023.6.0
ftfy==6.1.1
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.34
gradio==4.19.0
gradio_client==0.10.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
identify==2.5.33
idna==3.4
imageio==2.31.3
importlib-metadata==6.8.0
importlib-resources==6.0.1
iopath==0.1.10
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.17.3
jupyter-client==7.3.4
jupyter_core==5.3.1
jupyter-server==1.24.0
jupyterlab-pygments==0.2.2
kaggle==1.6.5
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.0.1
ml-collections==0.1.1
ml-dtypes==0.3.2
msgpack==1.0.7
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.25.2
omegaconf==2.3.0
openai==1.12.0
opencv-python==4.9.0.80
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
opt-einsum==3.3.0
optax==0.1.1
orbax-checkpoint==0.5.3
orjson==3.9.14
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pip==23.2.1
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.18.0
portalocker==2.8.2
pre-commit==3.6.0
preshed==3.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
python-magic==0.4.27
python-multipart==0.0.9
python-slugify==8.0.3
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==23.2.0
regex==2023.8.8
requests==2.31.0
rich==13.7.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
ruff==0.2.1
s3transfer==0.6.2
sacremoses==0.0.53
salesforce-lavis==1.0.2
scikit-image==0.21.0
scipy==1.11.2
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.30.0
setproctitle==1.3.2
setuptools==68.0.0
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.2
stanza==1.5.0
starlette==0.36.3
streamlit==1.31.0
tenacity==8.2.3
tensorflow-probability==0.23.0
tensorstore==0.1.53
terminado==0.17.1
text-unidecode==1.3
thinc==8.2.2
tifffile==2023.9.18
timm==0.4.12
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==1.10.0+cu111
torchaudio==0.10.0+rocm4.1
torchvision==0.11.0+cu111
tornado==6.1
tqdm==4.66.1
traitlets==5.9.0
transformers==4.26.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.16
uvicorn==0.27.1
validators==0.22.0
virtualenv==20.25.0
vqgan-jax==0.0.1
wandb==0.15.9
wasabi==1.1.2
watchdog==4.0.0
wcwidth==0.2.6
weasel==0.3.4
webdataset==0.2.86
webencodings==0.5.1
websocket-client==1.6.4
websockets==11.0.3
wheel==0.38.4
xxhash==3.4.1
yarl==1.9.4
zipp==3.16.2
